+++ gembackend/src/main/java/com/srilankagem/gembackend/gem/dto/GemStoneResponse.java
@@ -5,6 +5,7 @@ import com.srilankagem.gembackend.gem.models.GemTreatment;
 import com.srilankagem.gembackend.gem.models.GemType;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
@@ -18,10 +19,10 @@ public class GemStoneResponse {
     private String gemCode;
     private GemType type;
     private String color;
-    private Double caratWeight;
+    private BigDecimal caratWeight;
     private GemOrigin origin;
     private GemTreatment treatment;
-    private Double pricePerCarat;
+    private BigDecimal pricePerCarat;
     private Integer stockQuantity;
     private boolean certified;
     private String description;
